tutorsitelinks/requirements.txt
asgiref==3.3.1
autopep8==1.5.5
blis==0.4.1
catalogue==1.0.0
certifi==2020.12.5
chardet==4.0.0
cymem==2.0.3
dj-database-url==0.5.0
Django==3.1.7
djangorestframework==3.12.2
flake8==3.8.4
gunicorn==20.0.4
idna==2.10
importlib-metadata==3.7.0
mccabe==0.6.1
murmurhash==1.0.2
plac==1.1.3
preshed==3.0.2
psycopg2==2.8.6
pycodestyle==2.6.0
pyflakes==2.2.0
python-decouple==3.4
pytz==2021.1
requests==2.25.1
spacy==2.3.2
sqlparse==0.4.1
srsly==1.0.2
thinc==7.4.1
toml==0.10.2
typing-extensions==3.7.4.3
urllib3==1.26.4
wasabi==0.7.1
whitenoise==5.2.0
wincertstore==0.2
zipp==3.4.1
